Bump to v0.1.10 with Python SDK parity fixes (#1)
* Bump to v0.1.10 with Python SDK parity fixes

Squashed commits from feature/parity-sync:

- Add env-driven initialize timeout tests (CLAUDE_INITIALIZE_TIMEOUT)
- Support float values for HookMatcher timeout field
- Align version to 0.1.10 matching Python SDK

Test coverage: 2/2 passing (conformance + integration)

* Add GitHub Actions CI workflow

Build-only CI (no tests) since Claude CLI is not available in CI environment.
Builds on Ubuntu, Windows, and macOS in both Debug and Release configurations.

+++ include/claude/types.hpp
@@ -237,11 +237,11 @@ struct HookMatcher
     /// List of callback functions to invoke when hook matches
     std::vector<HookCallback> hooks;
 
-    /// Timeout in seconds for hook execution (default: 60)
-    std::optional<int> timeout;
+    /// Timeout in seconds for hook execution (default: 60). Accepts fractional seconds.
+    std::optional<double> timeout;
 
     HookMatcher() = default;
-    HookMatcher(std::optional<std::string> m, std::vector<HookCallback> h, std::optional<int> t = std::nullopt)
+    HookMatcher(std::optional<std::string> m, std::vector<HookCallback> h, std::optional<double> t = std::nullopt)
         : matcher(std::move(m)), hooks(std::move(h)), timeout(t) {}
 };
